+++ db/migrate/20170913151258_add_signature_to_users.rb
@@ -1,0 +1,5 @@
+class AddSignatureToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :signature, :text
+  end
+end