+++ test/application_system_test_case.rb
@@ -25,5 +25,6 @@ class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
 
   def complete_profile!(user)
     user.update(date_of_birth: "1984-12-31", address: "123 Road Way, City, ST 12345")
+    user.save_signature!(IO.readlines(File.join("test", "support", "signatures", "data_uri.txt")).first)
   end
 end